+++ src/components/home/contents/image-carousel/image-carousel.tsx
